requirements.txt
absl-py==2.3.0
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1749017159514/work
argon2-cffi-bindings @ file:///Users/runner/miniforge3/conda-bld/argon2-cffi-bindings_1762509511888/work
astunparse==1.6.3
attrs==25.3.0
bcrypt @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_bcrypt_1762497792/work
billiard==4.2.4
blinker==1.9.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1733827268945/work
branca==0.8.2
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
CacheControl==0.14.3
cachetools==5.5.2
ccxt==4.5.31
celery==5.6.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560578465/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coincurve==21.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1754405355057/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1742219570693/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1741620732069/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1764805067691/work
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
firebase-admin==6.4.0
Flask==3.1.1
Flask-Cors==4.0.1
flatbuffers==25.2.10
folium==0.20.0
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841362247/work
frozenlist==1.8.0
gast==0.6.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py==3.13.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302931775/work
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.2
kombu==5.6.2
lazy_loader==0.4
libclang==18.1.1
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1746516032751/work/libmambapy
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1731147022576/work
mpmath==1.3.0
msgpack==1.1.1
multidict==6.7.0
namex==0.1.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
onnxruntime==1.22.1
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1755779027131/work
passlib==1.7.4
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1733222765875/work
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.11.0
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1732588505673/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.1
PyJWT==2.10.1
PyMatting==1.1.14
pyparsing==3.2.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-bcrypt==0.3.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.21
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
redis==7.1.0
referencing==0.36.2
rembg==2.0.67
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1736248058782/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724504988/work
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.1
setuptools==80.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
SQLAlchemy==2.0.43
starlette==0.50.0
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1739009763/work
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
tzlocal==5.3.1
uritemplate==4.2.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.0
wheel==0.45.1
wrapt==1.17.2
xyzservices==2025.11.0
yarl==1.22.0
zstandard==0.23.0
